backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
comm==0.2.2
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
debugpy==1.8.12
decorator==5.1.1
Django==5.1.4
django-cors-headers==4.6.0
django-ninja==1.3.0
executing==2.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
pymeshlab==2023.12.post2
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
zope.interface==7.2
